requirements.txt
Flask==0.10.1
Flask-Login==0.2.6
Flask-SQLAlchemy==1.0
Flask-WTF==0.8.4
GitPython==0.3.2.RC1
Jinja2==2.7
MarkupSafe==0.18
SQLAlchemy==0.8.2
WTForms==1.0.4
Werkzeug==0.9.2
argparse==1.2.1
async==0.6.1
gitdb==0.5.4
github3.py==0.7.0
gunicorn==17.5
itsdangerous==0.22
requests==1.2.3
smmap==0.8.2
wsgiref==0.1.2
